% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_functions.R
\name{cats_over_dogs}
\alias{cats_over_dogs}
\title{A Cat Function}
\usage{
cats_over_dogs(agree = TRUE)
}
\arguments{
\item{agree}{Do you agree cats are the best pet? Defaults to TRUE.}
}
\description{
This function allows you to express your love for the superior furry animal.
}
\examples{
cats_over_dogs()
}
\keyword{cats}
