% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cso_search_url.R
\name{create_cso_search_url}
\alias{create_cso_search_url}
\title{Convenience function to generate a correct StatBank search URL}
\usage{
create_cso_search_url(search_term, offset)
}
\arguments{
\item{search_term}{A search term used to get CSO datasets.}

\item{offset}{The offset at which to retrieve the subsequent 15 dataset names.}
}
\value{
A valid URL for searching StatBank datasets.
}
\description{
Convenience function to generate a correct StatBank search URL
}
\keyword{internal}
