% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataKindR-package.R
\docType{package}
\name{DataKindR-package}
\alias{DataKindR}
\alias{DataKindR-package}
\title{DataKindR: Provides Helper Functions for DataKind Volunteers}
\description{
DataKind volunteers often need access to specific data or techniques in a DataDive or other project. This package seeks to make this access easier.
}
\author{
\strong{Maintainer}: Robin Penfold \email{robin.penfold@icloud.com}

}
\keyword{internal}
