% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datakindr.R
\docType{package}
\name{datakindr}
\alias{datakindr}
\alias{datakindr-package}
\title{datakindr: Tools And Utilities for DataKind}
\description{
Provides tools, utilities, data and more for use by the DataKind
Dublin team. Including a convenient interface to CSO.ie's Statbank and a
ggplot2 theme and colour palette for convenient & consistent visualisation
styles.
}
\section{Datakindr statbank functions}{

For interacting with CSO.ie datasets, the following two functions can be used:
\itemize{
\item{\code{search_statbank_datasets()} returns a data frame with dataset
codes and a description of the dataset to which each code refers.}
\item{Once you've found the dataset you need (either from the Statbank website, or
by using \code{search_statbank_datasets()}) you can use
\code{get_cso_dataset()} to retrieve the dataset as a data.frame.}
}
}

\section{Datakindr Plotting}{

Blah blah blah
}

