% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datakind_palette.R
\docType{data}
\name{dk_palette}
\alias{dk_palette}
\title{DataKind Recommended Palette}
\format{A list of strings, each is a colour hex code.
\describe{
  \item{dk_palette}{list of datakind colour hex codes}
}}
\usage{
dk_palette
}
\description{
A list containing a the recommended palette for use when making visualisations
for DataKind.
}
\keyword{datasets}
