% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_sequential_palette.R
\docType{data}
\name{dk_sequential_palette}
\alias{dk_sequential_palette}
\title{DataKind Recommended Palette For Sequential Data}
\format{A list of strings, each is a colour hex code.
\describe{
  \item{dk_palette}{List of datakind colour hex codes.}
  \item{colours:}{("#404040", "#804d00" , "#BB6600" , "#FF6600", "#FF9900", "#FFCC33" , "#FFCC99")}
}}
\usage{
dk_sequential_palette
}
\description{
A list containing an extended palette for use when making visualisations
for DataKind with data in sequential order.
}
\keyword{datasets}
