% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cso_dataset.R
\name{get_cso_dataset}
\alias{get_cso_dataset}
\title{Gets a CSO.ie StatBank dataset when given a dataset code.}
\usage{
get_cso_dataset(cso_dataset_code)
}
\arguments{
\item{cso_dataset_code}{A dataset code.}
}
\value{
A data frame of arbitrary dimension and size containing a StatBank dataset.
}
\description{
Gets a CSO.ie StatBank dataset when given a dataset code.
}
\examples{
get_cso_dataset("LRM07")
}
