% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_search_results.R
\name{get_num_search_results}
\alias{get_num_search_results}
\title{Gets the number of search results that a given search term on StatBank
will return. Uses web scraping. For internal use by datakindr.}
\usage{
get_num_search_results(search_term)
}
\arguments{
\item{search_term}{A search term used to get CSO datasets.}
}
\value{
The number of search results
}
\description{
Gets the number of search results that a given search term on StatBank
will return. Uses web scraping. For internal use by datakindr.
}
\keyword{internal}
