% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_node_text.R
\name{parse_node_text}
\alias{parse_node_text}
\title{Parse HTML Node Text}
\usage{
parse_node_text(xml_doc, node_name)
}
\arguments{
\item{xml_doc}{xmldocument object for which you want a node parsed and
returned as text.}

\item{node_name}{Name of the node from which you wish to extract text.}
}
\value{
A list with each line of text that rvest parses
}
\description{
Uses rvest to extract the text from a node into a list.
}
\keyword{internal}
