% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results_page.R
\name{parse_results_page}
\alias{parse_results_page}
\title{Title}
\usage{
parse_results_page(httr_response)
}
\arguments{
\item{httr_response}{A httr response object from a Statbank page. Can
either be a search results page or a dataset page (i.e., only one
result was returned).}
}
\value{
Returns a list of datasets codes and names as a single string per
dataset.
}
\description{
Title
}
\keyword{internal}
