% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{postcodes}
\alias{postcodes}
\title{Clean UK postcodes and extract their relevant elements}
\usage{
postcodes(postcode_value, postcode_type = "full")
}
\arguments{
\item{postcode_value}{The postcode data to be cleaned (in the form of 'N1 1AA', 'ME1 2RE' or 'TN12 0QS').}

\item{postcode_type}{The relevant section of the postcode to return. Only one section is permitted. Options are:
\describe{
  \item{full}{the full postcode (such as 'N1 1AA')}
  \item{region}{the first part of the first element of the postcode (such as 'ME' for 'ME1 2RE')}
  \item{area}{the first element of the postcode (such as 'ME1' for 'ME1 2RE')}
  \item{locale}{the area and the first part of the second element of the postcode (such as 'ME1 2' for 'ME1 2RE')}
  }}
}
\description{
This function cleans UK postcode data before returning the relevant section (such as its first part of full postcode).
}
\examples{
postcodes(
  postcode_value = 'ME1 2RE',
  postcode_type = 'full'
  )

}
\keyword{Postcodes}
