% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes_metadata.R
\name{postcodes_metadata}
\alias{postcodes_metadata}
\title{Scrape geographic metadata about UK postcodes}
\usage{
postcodes_metadata(postcode_value)
}
\arguments{
\item{postcode_value}{The postcode data to be cleaned (in the form of 'N1 1AA', 'ME1 2RE' or 'TN12 0QS'). To clean this data, you could pass it through the `postcodes` function in this package.}
}
\description{
This function takes clean UK postcode data and returns key geographic metadata. Details include: LSOA, MSOA, Latitude, Longitude, Deprivation Index and Constituency.
}
\examples{
postcodes_metadata(
  postcode_value = "ME1 2RE"
  )
}
\keyword{Postcodes}
