% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results_to_dataframe.R
\name{search_results_to_dataframe}
\alias{search_results_to_dataframe}
\title{Takes the raw output list from the web scraping of the Statbanks search
page and returns a data frame. For internal use of the package.}
\usage{
search_results_to_dataframe(search_results)
}
\arguments{
\item{request}{A completed httr request.}
}
\value{
The inner HTML content of the httr request.
}
\description{
Takes the raw output list from the web scraping of the Statbanks search
page and returns a data frame. For internal use of the package.
}
\keyword{internal}
