% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_get_num_results.R
\name{try_get_num_results}
\alias{try_get_num_results}
\title{Try To Get The Number of Search Results}
\usage{
try_get_num_results(res)
}
\arguments{
\item{res}{}
}
\value{
The number of search results from the Statbank webpage.
}
\description{
Uses a \code{tryCatch()} block to get the number of search results, and
raises a warning when no results returned.
}
\keyword{internal}
