% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_search_term.R
\name{validate_search_term}
\alias{validate_search_term}
\title{Ensures the search results are useful to the user. Throws errors if
validation fails.}
\usage{
validate_search_term(search_term, num_results, max_num_results)
}
\arguments{
\item{search_term}{A search term used to get CSO datasets.}

\item{num_results}{The number of results for the current search term.}

\item{max_num_results}{The maximum number of results to return.}
}
\description{
Ensures the search results are useful to the user. Throws errors if
validation fails.
}
\keyword{internal}
