% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataKindR-package.R
\docType{package}
\name{DataKindR-package}
\alias{DataKindR}
\alias{DataKindR-package}
\title{DataKindR: Provides Helper Functions for DataKind Volunteers}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

DataKind volunteers often need access to specific data or techniques in a DataDive or other project. This package seeks to simplify access to these resources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/p0bs/DataKindR}
  \item Report bugs at \url{https://github.com/p0bs/DataKindR/issues}
}

}
\author{
\strong{Maintainer}: Robin Penfold \email{robin.penfold@icloud.com}

}
\keyword{internal}
