% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datakind_palette.R
\docType{data}
\name{dk_palette}
\alias{dk_palette}
\title{DataKind Recommended Palette}
\format{A list of strings, each is a colour hex code.
\describe{
  \item{dk_palette}{List of datakind colour hex codes.}
  \item{colours:}{("#404040", "#FF6600", "#804D00", "#FF9900", "#FFFFFF", "#EEEEDD")}
}}
\usage{
dk_palette
}
\description{
A list containing a the recommended palette for use when making visualisations
for DataKind.
}
\keyword{datasets}
