% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_httr_html_content.R
\name{get_httr_html_content}
\alias{get_httr_html_content}
\title{Internal convenience wrapper to get the inner HTML content given a
valid URL request.}
\usage{
get_httr_html_content(request)
}
\arguments{
\item{request}{A completed httr request.}
}
\value{
The inner HTML content of the httr request.
}
\description{
Internal convenience wrapper to get the inner HTML content given a
valid URL request.
}
\keyword{internal}
