% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_statbank_datasets.R
\name{search_statbank_datasets}
\alias{search_statbank_datasets}
\title{Retrives a list of CSO Statbank dataset descriptions and dataset codes.
These codes can be used with get_cso_dataset in this package.}
\usage{
search_statbank_datasets(search_term, max_num_results = 500)
}
\arguments{
\item{search_term}{A search term used to get CSO datasets.}

\item{max_num_results}{The maximum number of results to return.}
}
\value{
A data frame with two columns, \code{dataset_code} &
\code{dataset_desc}. Codes can be used to retrieve Statbank datasets.
}
\description{
Retrives a list of CSO Statbank dataset descriptions and dataset codes.
These codes can be used with get_cso_dataset in this package.
}
\examples{
search_statbank_datasets("employment 2002")
search_statbank_datasets("age", 1001)
}
